update navigation to alert controller
