+++ AppVTB/AppVTB/FavoritesScreen/Router/FavoritesRouter.swift
@@ -10,6 +10,15 @@ import UIKit
 
 final class FavoritesRouter {
     
+    //MARK: - Constants
+    
+    enum Locals {
+        static let title = "Add new data"
+        static let addTitle = "Add"
+        static let cancelTitle = "Cancel"
+        static let placeholder = "Enter email or phone number"
+    }
+    
     // MARK: - Properties
     
     unowned let view: UIViewController
@@ -21,9 +30,20 @@ final class FavoritesRouter {
 
 extension FavoritesRouter: FavoritesRouterInput {
     func showAlert() {
-        let alertController = FavoritesAlertBuilder.createAlert()
+        let alertController = UIAlertController(title: Locals.title, message: nil, preferredStyle: .alert)
+        let confirmAction = UIAlertAction(title: Locals.addTitle, style: .default) { (_) in
+            if let txtField = alertController.textFields?.first, let text = txtField.text, !text.isEmpty {
+                //do stuff
+            }
+        }
+        alertController.addAction(confirmAction)
+        let cancelAction = UIAlertAction(title: Locals.cancelTitle, style: .cancel) { (_) in }
+        alertController.addAction(cancelAction)
+        alertController.addTextField { (textField) in
+            textField.placeholder = Locals.placeholder
+        }
         view.present(alertController, animated: true, completion: nil)
     }
     
-
+    
 }
